rename: handler, response, exception 패키지 이동 (domain/common/ -> global/)
