+++ src/main/java/com/example/springrider/domain/cart/enums/CartItemStatus.java
@@ -1,7 +1,7 @@
 package com.example.springrider.domain.cart.enums;
 
-import com.example.springrider.domain.common.exception.ExceptionCode;
-import com.example.springrider.domain.common.exception.InvalidRequestException;
+import com.example.springrider.global.exception.ExceptionCode;
+import com.example.springrider.global.exception.InvalidRequestException;
 import java.util.Arrays;
 
 public enum CartItemStatus {
