+++ src/main/java/com/example/springrider/domain/cart/service/CartService.java
@@ -12,8 +12,8 @@ import com.example.springrider.domain.cart.dto.UpdateCartItemResponseDto;
 import com.example.springrider.domain.cart.entity.CartItem;
 import com.example.springrider.domain.cart.enums.CartItemStatus;
 import com.example.springrider.domain.cart.repository.CartRepository;
-import com.example.springrider.domain.common.exception.ExceptionCode;
-import com.example.springrider.domain.common.exception.InvalidRequestException;
+import com.example.springrider.global.exception.ExceptionCode;
+import com.example.springrider.global.exception.InvalidRequestException;
 import com.example.springrider.domain.menu.entity.Menu;
 import com.example.springrider.domain.menu.repository.MenuRepository;
 import com.example.springrider.domain.user.entity.User;
