+++ src/main/java/com/example/springrider/domain/menu/repository/MenuRepository.java
@@ -1,7 +1,7 @@
 package com.example.springrider.domain.menu.repository;
 
-import com.example.springrider.domain.common.exception.ExceptionCode;
-import com.example.springrider.domain.common.exception.InvalidRequestException;
+import com.example.springrider.global.exception.ExceptionCode;
+import com.example.springrider.global.exception.InvalidRequestException;
 import com.example.springrider.domain.menu.entity.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
