+++ src/main/java/com/example/springrider/domain/order/controller/UserOrderController.java
@@ -1,6 +1,6 @@
 package com.example.springrider.domain.order.controller;
 
-import com.example.springrider.domain.common.response.ApiResponse;
+import com.example.springrider.global.response.ApiResponse;
 import com.example.springrider.domain.order.dto.request.CreateOrderRequestDto;
 import com.example.springrider.domain.order.dto.response.OrderResponseDto;
 import com.example.springrider.domain.order.service.UserOrderService;
