+++ src/main/java/com/example/springrider/domain/order/service/UserOrderService.java
@@ -3,8 +3,8 @@ package com.example.springrider.domain.order.service;
 import com.example.springrider.aop.OrderEventLog;
 import com.example.springrider.domain.cart.entity.CartItem;
 import com.example.springrider.domain.cart.repository.CartRepository;
-import com.example.springrider.domain.common.exception.ExceptionCode;
-import com.example.springrider.domain.common.exception.InvalidRequestException;
+import com.example.springrider.global.exception.ExceptionCode;
+import com.example.springrider.global.exception.InvalidRequestException;
 import com.example.springrider.domain.order.dto.request.CreateOrderRequestDto;
 import com.example.springrider.domain.order.dto.response.OrderResponseDto;
 import com.example.springrider.domain.order.entity.Order;
