+++ src/main/java/com/example/springrider/global/exception/ExceptionCode.java
@@ -1,4 +1,4 @@
-package com.example.springrider.domain.common.exception;
+package com.example.springrider.global.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
